update deep netwrok and adjust the functins
